requirements.txt
PyYAML
requests
prometheus-client
python-dateutil